fixes ready
