+++ controllers/controllerNews.js
@@ -1,28 +1,19 @@
 const serviceNews = require('../services/news');
-const { check, validationResult, body } = require('express-validator');
 
-const controllerNews = async function (req, res, next) {
-    try {
-        await check('name').notEmpty().run(req);
-        await check('image').notEmpty().run(req);
-        await check('content').notEmpty().run(req);
-        //await check('categoryId').notEmpty().run(req);
-        const result = validationResult(req);
-        if (!result.isEmpty()) {
-            return res.status(400).json({ errors: result.array() });
-        }
+const controllerAddNews = async function (req, res, next) {
+    try {const { name, content, image } = req
         const data = {
-            name: req.body.name,
-            content: req.body.content,
-            image: req.body.image
+            name,
+            content,
+            image
         }
-        data1 = await serviceNews.save(data);
-        res.status(201).json('the news were successfully saved ')
-        return data1
+        const newsCreated = await serviceNews.save(data);
+        res.status(201).json('The news were successfully saved ')
+        return newsCreated;
     } catch (error) {
         next(error)
         throw new Error("something happened")
     }
-}
+};
 
-module.exports = { controllerNews };
+module.exports = { controllerAddNews };
