+++ repositories/news.js
@@ -1,19 +1,13 @@
 const DbNews = require('../models');
 
-class newsRepository {
-
-    constructor() {
-
+async function addNews(news) {
+    try {
+        const novedades = await DbNews.New.create(news);
+        return novedades;
+    } catch (error) {
+        throw new Error('Wrong information');
     }
 
-    async addNews(news) {
-        try {
-            const novedades = await DbNews.New.create(news);
-            return novedades;
-        } catch (error) {
-            throw new Error('Wrong information');
-        }
+}
 
-    }
-};
-module.exports = newsRepository;
+module.exports = addNews;
