+++ services/news.js
@@ -1,7 +1,6 @@
 const addNews = require('../repositories/news');
-const newsRepo = new addNews();
 
 const save = async(news) =>{
-    return await newsRepo.addNews(news);
+    return await addNews(news);
 }
 module.exports = {save};
